Chore: re-organise Root, add folder for SubHeader
