+++ src/Root.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import App from '../containers/App';
-import store from '../store/configureStore';
+import App from './containers/App';
+import store from './store/configureStore';
 
 const Root = () => (
   <Provider store={store}>
