+++ src/main.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import configureStore from './store/configureStore';
-import Root from './components/Root';
+import Root from './Root';
 
 const store = configureStore();
