+++ src/styles/index.css
@@ -22,7 +22,7 @@ html {
 }
 
 .app-main {
-  display: flex;
+  display: block;
   min-width: 100%;
   min-height: 100vh;
 }
